chg: update to SharpDevelop 2.1 final (build 2.1.0.2429)
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%40561

+++ Libraries/SharpDevelop.Base/Project/Src/Gui/Components/FontSelectionPanel.cs
@@ -2,7 +2,7 @@
 //     <copyright see="prj:///doc/copyright.txt"/>
 //     <license see="prj:///doc/license.txt"/>
 //     <owner name="none" email=""/>
-//     <version>$Revision: 1965 $</version>
+//     <version>$Revision: 2273 $</version>
 // </file>
 
 using System;
@@ -199,9 +199,9 @@ namespace ICSharpCode.SharpDevelop.Gui
 		{
 			if (!fontListComboBox.Enabled)
 				return null;
-			int fontSize = 10;
+			float fontSize = 10f;
 			try {
-				fontSize = Math.Max(6, Int32.Parse(fontSizeComboBox.Text));
+				fontSize = Math.Max(6, Single.Parse(fontSizeComboBox.Text));
 			} catch (Exception) {}
 			
 			FontDescriptor fontDescriptor = (FontDescriptor)fontListComboBox.Items[fontListComboBox.SelectedIndex];
